# api/openapi.yaml
openapi: 3.0.3

info:
  title: Avito Tamagotchi API
  version: 1.0.0
  description: Authentication, game, and WebSocket contract for the Avito Tamagotchi MVP.

servers:
  - url: /api/v1
    description: HTTP API through Caddy

security:
  - sessionCookie: []

paths:
  /auth/register:
    post:
      operationId: registerUser
      summary: Register a user and create an authenticated session
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User registered and authenticated
          headers:
            Set-Cookie:
              $ref: "#/components/headers/SessionCookie"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/ValidationError"
        "409":
          $ref: "#/components/responses/EmailAlreadyExistsError"
        "500":
          $ref: "#/components/responses/InternalServerError"

  /auth/login:
    post:
      operationId: loginUser
      summary: Log in and create an authenticated session
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: User authenticated
          headers:
            Set-Cookie:
              $ref: "#/components/headers/SessionCookie"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/InvalidCredentialsError"
        "500":
          $ref: "#/components/responses/InternalServerError"

  /auth/logout:
    post:
      operationId: logoutUser
      summary: Delete the current session and clear the session cookie
      security: []
      responses:
        "204":
          description: Session deleted when a valid session cookie is present; cookie is cleared in all cases
          headers:
            Set-Cookie:
              $ref: "#/components/headers/ExpiredSessionCookie"
        "500":
          $ref: "#/components/responses/InternalServerError"

  /auth/me:
    get:
      operationId: getCurrentUser
      summary: Get the authenticated user
      responses:
        "200":
          description: Current authenticated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "500":
          $ref: "#/components/responses/InternalServerError"

  /ws:
    get:
      operationId: connectWebSocket
      summary: Connect to authenticated real-time events
      description: >-
        Upgrades the authenticated HTTP request to a WebSocket connection.
        The server sends JSON event envelopes with an event name and does not
        send an initial snapshot. Browser handshakes must use an Origin whose
        host matches the request Host; non-browser clients may omit Origin.
        The server sends ping frames every 54 seconds; each pong extends the
        read deadline by 60 seconds. Client messages are ignored, but each
        message is limited to 512 bytes.
      servers:
        - url: /
          description: WebSocket endpoint through Caddy
      x-websocket-messages:
        $ref: "#/components/schemas/WSEvent"
      responses:
        "101":
          description: WebSocket connection established
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          description: The WebSocket Origin does not match the request Host.
          content:
            text/plain:
              schema:
                type: string
                example: Forbidden

  /pet:
    get:
      operationId: getPet
      summary: Get or create the current user's pet
      responses:
        "200": { description: Pet stats, content: { application/json: { schema: { $ref: "#/components/schemas/Pet" } } } }
        "401": { $ref: "#/components/responses/UnauthorizedError" }
        "500": { $ref: "#/components/responses/InternalServerError" }
  /pet/actions:
    post:
      operationId: chargePet
      summary: Charge the pet
      description: >-
        The first charge in a Moscow calendar day restores energy to 100 and
        grants 10 base XP plus a streak-based daily reward from 10 to 50 XP.
        Later calls on the same day return zero awarded XP.
      requestBody:
        required: true
        content:
          application/json: { schema: { type: string, enum: [charge] } }
      responses:
        "200": { description: Updated pet and awarded XP, content: { application/json: { schema: { $ref: "#/components/schemas/ChargeResult" } } } }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/UnauthorizedError" }
        "500": { $ref: "#/components/responses/InternalServerError" }
  /streak:
    get:
      operationId: getStreak
      summary: Get the current user's charge streak
      responses:
        "200": { description: Current and longest streak, content: { application/json: { schema: { $ref: "#/components/schemas/Streak" } } } }
        "401": { $ref: "#/components/responses/UnauthorizedError" }
        "500": { $ref: "#/components/responses/InternalServerError" }
  /tasks/today:
    get:
      operationId: getTodayTasks
      summary: Get today's rotated tasks
      responses:
        "200": { description: Tasks, content: { application/json: { schema: { $ref: "#/components/schemas/Tasks" } } } }
        "401": { $ref: "#/components/responses/UnauthorizedError" }
        "500": { $ref: "#/components/responses/InternalServerError" }
  /mock-avito/actions:
    post:
      operationId: applyMockAvitoAction
      summary: Apply a mock Avito action
      requestBody:
        required: true
        content:
          application/json: { schema: { type: string, enum: [view, favorite, create_listing, create_listing_in_category] } }
      responses:
        "200": { description: Action accepted, content: { application/json: { schema: { $ref: "#/components/schemas/ActionResult" } } } }
        "400":
          description: The action is invalid or is not available today.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401": { $ref: "#/components/responses/UnauthorizedError" }
        "409": { $ref: "#/components/responses/PetDeadError" }
        "500": { $ref: "#/components/responses/InternalServerError" }
  /rewards:
    get:
      operationId: getRewards
      summary: Get unlocked rewards
      responses:
        "200": { description: Rewards, content: { application/json: { schema: { $ref: "#/components/schemas/Rewards" } } } }
        "401": { $ref: "#/components/responses/UnauthorizedError" }
        "500": { $ref: "#/components/responses/InternalServerError" }
  /rewards/{rewardID}/use:
    post:
      operationId: useReward
      summary: Use an unlocked reward
      parameters: [{ name: rewardID, in: path, required: true, schema: { type: string, format: uuid } }]
      responses:
        "200": { description: Used reward, content: { application/json: { schema: { $ref: "#/components/schemas/Reward" } } } }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/UnauthorizedError" }
        "404": { $ref: "#/components/responses/RewardNotFoundError" }
        "409": { $ref: "#/components/responses/RewardAlreadyUsedError" }
        "500": { $ref: "#/components/responses/InternalServerError" }
  /summary/today:
    get:
      operationId: getTodaySummary
      summary: Get today's Moscow-time summary
      responses:
        "200": { description: Daily summary, content: { application/json: { schema: { $ref: "#/components/schemas/Summary" } } } }
        "401": { $ref: "#/components/responses/UnauthorizedError" }
        "500": { $ref: "#/components/responses/InternalServerError" }
  /leaderboard:
    get:
      operationId: getLeaderboard
      summary: Get the all-time XP top ten and current user's rank
      responses:
        "200": { description: Leaderboard, content: { application/json: { schema: { $ref: "#/components/schemas/Leaderboard" } } } }
        "401": { $ref: "#/components/responses/UnauthorizedError" }
        "500": { $ref: "#/components/responses/InternalServerError" }
  /leaderboard/week:
    get:
      operationId: getWeeklyLeaderboard
      summary: Get the XP top ten for the last 7 times 24 hours
      responses:
        "200": { description: Weekly XP delta leaderboard, content: { application/json: { schema: { $ref: "#/components/schemas/DeltaLeaderboard" } } } }
        "401": { $ref: "#/components/responses/UnauthorizedError" }
        "500": { $ref: "#/components/responses/InternalServerError" }
  /leaderboard/month:
    get:
      operationId: getMonthlyLeaderboard
      summary: Get the XP top ten for the last 30 times 24 hours
      responses:
        "200": { description: Monthly XP delta leaderboard, content: { application/json: { schema: { $ref: "#/components/schemas/DeltaLeaderboard" } } } }
        "401": { $ref: "#/components/responses/UnauthorizedError" }
        "500": { $ref: "#/components/responses/InternalServerError" }
  /leaderboard/streak:
    get:
      operationId: getStreakLeaderboard
      summary: Get the active charge streak top ten
      responses:
        "200": { description: Active streak leaderboard, content: { application/json: { schema: { $ref: "#/components/schemas/StreakLeaderboard" } } } }
        "401": { $ref: "#/components/responses/UnauthorizedError" }
        "500": { $ref: "#/components/responses/InternalServerError" }

components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session_id
      description: HttpOnly session cookie with a seven-day lifetime.

  headers:
    SessionCookie:
      description: >-
        Session cookie set on successful registration or login. Uses Path=/,
        HttpOnly, SameSite=Lax, a seven-day expiration, and Secure when
        APP_SECURE_COOKIES=true.
      schema:
        type: string

    ExpiredSessionCookie:
      description: Session cookie cleared with Max-Age=-1 and an expiration date in the past.
      schema:
        type: string

  schemas:
    RegisterRequest:
      type: object
      additionalProperties: false
      required:
        - email
        - password
        - displayName
      properties:
        email:
          type: string
          format: email
          maxLength: 254
          description: Trimmed and normalized to lowercase by the backend.
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 128
        displayName:
          type: string
          minLength: 2
          maxLength: 32
          pattern: "^[A-Za-z_-]+$"

    LoginRequest:
      type: object
      additionalProperties: false
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          maxLength: 254
          description: Trimmed and normalized to lowercase by the backend.
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 128

    User:
      type: object
      additionalProperties: false
      required:
        - id
        - email
        - displayName
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        displayName:
          type: string
        createdAt:
          type: string
          format: date-time

    WSEvent:
      type: object
      additionalProperties: false
      required:
        - event
      properties:
        event:
          type: string
          enum: [pet_updated, tasks_updated, rewards_updated]
          description: Event name published by the backend.

    Pet:
      type: object
      additionalProperties: false
      required: [xp, level, stage, energy, lastChargedAt, isDead]
      properties:
        xp: { type: integer, minimum: 0 }
        level: { type: integer, minimum: 1 }
        stage: { type: string, enum: [egg, child, teen, adult] }
        energy: { type: integer, minimum: 0, maximum: 100 }
        lastChargedAt: { type: string, format: date-time }
        isDead: { type: boolean }
    ChargeResult:
      type: object
      additionalProperties: false
      required: [pet, baseChargeXp, dailyRewardXp, totalXpAwarded]
      properties:
        pet: { $ref: "#/components/schemas/Pet" }
        baseChargeXp: { type: integer, minimum: 0, maximum: 10 }
        dailyRewardXp: { type: integer, minimum: 0, maximum: 50 }
        totalXpAwarded: { type: integer, minimum: 0, maximum: 60 }
    Streak:
      type: object
      additionalProperties: false
      required: [currentStreak, longestStreak, lastChargeDate, nextDailyRewardXp]
      properties:
        currentStreak: { type: integer, minimum: 0 }
        longestStreak: { type: integer, minimum: 0 }
        lastChargeDate: { type: string, format: date, nullable: true }
        nextDailyRewardXp: { type: integer, minimum: 10, maximum: 50 }
    Tasks:
      type: object
      additionalProperties: false
      required: [tasks]
      properties:
        tasks: { type: array, items: { $ref: "#/components/schemas/Task" } }
    Task:
      type: object
      additionalProperties: false
      required: [type, progress, requiredCount, status, xpReward]
      properties:
        type: { type: string, enum: [view, favorite, create_listing, create_listing_in_category] }
        progress: { type: integer, minimum: 0 }
        requiredCount: { type: integer, minimum: 1 }
        status: { type: string, enum: [in_progress, completed] }
        xpReward: { type: integer, minimum: 1 }
    ActionResult:
      type: object
      additionalProperties: false
      required: [status]
      properties:
        status: { type: string, enum: [applied] }
    Rewards:
      type: object
      additionalProperties: false
      required: [rewards]
      properties:
        rewards:
          type: array
          items: { $ref: "#/components/schemas/Reward" }
    Reward:
      type: object
      additionalProperties: false
      required: [id, level, type, status, unlockedAt, usedAt]
      properties:
        id: { type: string, format: uuid }
        level: { type: integer, minimum: 2 }
        type: { type: string, enum: [promotion, free_delivery] }
        status: { type: string, enum: [available, used] }
        unlockedAt: { type: string, format: date-time }
        usedAt: { type: string, format: date-time, nullable: true }
    Summary:
      type: object
      additionalProperties: false
      required: [xpEarned, completedTasks, charges, level, currentXP, energy, unlockedRewards]
      properties:
        xpEarned: { type: integer, minimum: 0 }
        completedTasks: { type: integer, minimum: 0 }
        charges: { type: integer, minimum: 0 }
        level: { type: integer, minimum: 1 }
        currentXP: { type: integer, minimum: 0 }
        energy: { type: integer, minimum: 0, maximum: 100 }
        unlockedRewards: { type: array, items: { type: string, enum: [promotion, free_delivery] } }
    Leaderboard:
      type: object
      additionalProperties: false
      required: [entries, currentUser]
      properties:
        entries: { type: array, maxItems: 10, items: { $ref: "#/components/schemas/LeaderboardEntry" } }
        currentUser:
          allOf:
            - $ref: "#/components/schemas/LeaderboardEntry"
          nullable: true
    LeaderboardEntry:
      type: object
      additionalProperties: false
      required: [rank, userId, displayName, xp, level]
      properties:
        rank: { type: integer, minimum: 1 }
        userId: { type: string, format: uuid }
        displayName: { type: string }
        xp: { type: integer, minimum: 0 }
        level: { type: integer, minimum: 1 }
    DeltaLeaderboard:
      type: object
      additionalProperties: false
      required: [entries, currentUser]
      properties:
        entries: { type: array, maxItems: 10, items: { $ref: "#/components/schemas/DeltaLeaderboardEntry" } }
        currentUser:
          allOf:
            - $ref: "#/components/schemas/DeltaLeaderboardEntry"
          nullable: true
    DeltaLeaderboardEntry:
      type: object
      additionalProperties: false
      required: [rank, userId, displayName, xp, level, xpDelta]
      properties:
        rank: { type: integer, minimum: 1 }
        userId: { type: string, format: uuid }
        displayName: { type: string }
        xp: { type: integer, minimum: 0 }
        level: { type: integer, minimum: 1 }
        xpDelta: { type: integer, minimum: 0 }
    StreakLeaderboard:
      type: object
      additionalProperties: false
      required: [entries, currentUser]
      properties:
        entries: { type: array, maxItems: 10, items: { $ref: "#/components/schemas/StreakLeaderboardEntry" } }
        currentUser:
          allOf:
            - $ref: "#/components/schemas/StreakLeaderboardEntry"
          nullable: true
    StreakLeaderboardEntry:
      type: object
      additionalProperties: false
      required: [rank, userId, displayName, currentStreak, longestStreak, streakStartedAt, lastChargeDate]
      properties:
        rank: { type: integer, minimum: 1 }
        userId: { type: string, format: uuid }
        displayName: { type: string }
        currentStreak: { type: integer, minimum: 1 }
        longestStreak: { type: integer, minimum: 1 }
        streakStartedAt: { type: string, format: date }
        lastChargeDate: { type: string, format: date }

    ErrorResponse:
      type: object
      additionalProperties: false
      required:
        - code
        - message
      properties:
        code:
          type: string
          enum:
            - validation_error
            - unauthorized
            - invalid_credentials
            - email_already_exists
            - task_not_available
            - pet_dead
            - reward_not_found
            - reward_already_used
            - internal_error
        message:
          type: string

  responses:
    ValidationError:
      description: >-
        Request body is malformed, contains trailing JSON values or unknown
        fields, exceeds the endpoint limit, or fails backend validation.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

    UnauthorizedError:
      description: Authentication is required or the session has expired.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

    PetDeadError:
      description: The pet must be charged before task actions are available.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

    InvalidCredentialsError:
      description: Email or password is incorrect.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

    EmailAlreadyExistsError:
      description: The normalized email is already registered.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

    RewardNotFoundError:
      description: The reward does not exist or belongs to another user.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

    RewardAlreadyUsedError:
      description: The reward has already been used.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

    InternalServerError:
      description: Internal server error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"